fix(dht): remove unnecessary async in kademlia behaviour
