+++ market_dht/Cargo.toml
@@ -20,6 +20,7 @@ libp2p = { version = "0.53.2", features = [
   "yamux",
   "identify",
   "macros",
+  "request-response",
 ] }
 futures = { version = "0.3.30" }
 thiserror = { version = "1.0.58" }
